Mad Lib can now be completed by a user.
The _formStepper() method walks the user through all of the fields in the
Mad Lib, and once finished, the Lib() method replaces the jumbled version
with the completed version.

Currently this uses javascript prompts (gross!), but the next commit should
introduce a modal window to clean this up a bit.