Realtalk:
This is a hot mess.  The initial design was recursive with an expectation
that a javascript prompt was going to interrupt the return for each
function call.  However, since switching to a prettier dialog-based format,
the need has arisen for tracking the user's progress through the Mad Lib
and modifying the UI on that basis.

Because of this, there are now a number of unreliable SetTimeout and
SetInterval listeners that need to Go Away because they cause weird race
conditions and inconsistent behavior.  Essentially, the bolted-on status
monitoring code is inadequate for providing a robust Mad Lib solution
(lol), so a new approach and significant refactoring is required.