It ain't pretty, but...
It's working well enough.  Free-form Mad Libs can be sent into the applet
and will be processed by the dialog into completed versions.

Since this is a once-off for a prank website, the dialog cannot be closed
until the mad lib is completed.  Partially completed mad-libs are not
supported, though it would be fairly trivial to implement.